# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "L5 OpenApi",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "email": "darius@matulionis.lt"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0",
        "x-logo": {
            "url": "https://via.placeholder.com/190x90.png?text=L5-Swagger"
        }
    },
    "paths": {
...
        "/api/login": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "login to the account",
                "description": " ",
                "operationId": "login",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LoginRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "400": {
                        "description": "Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Error"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error in Requried Inputs",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Error Message",
                                            "type": "string",
                                            "example": "Unprocessable Entity"
                                        },
                                        "errors": {
                                            "description": "Errors",
                                            "type": "Object",
                                            "example": "{'email': ['الإيميل مطلوب'   ]}"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "logout from the account",
                "operationId": "logout",
                "responses": {
                    "400": {
                        "description": "Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Error"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                },
                                "example": "{data:logout}"
                            }
                        }
                    }
                }
            }
        },
        "/api/reset-password": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "  otp verify",
                "description": " ",
                "operationId": "password_reset_otp_verify",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/OtpVerifyRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "400": {
                        "description": "Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Error"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error in Requried Inputs",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Error Message",
                                            "type": "string",
                                            "example": "Unprocessable Entity"
                                        },
                                        "errors": {
                                            "description": "Errors",
                                            "type": "Object",
                                            "example": "{'email': ['الإيميل مطلوب'   ]}"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "oneOf": [
                                                {
                                                    "$ref": "#/components/schemas/User"
                                                }
                                            ]
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/forgot-password": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "password  reset otp send",
                "description": " ",
                "operationId": "password_reset_otp_send",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ForgetPasswordOtpSendRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "400": {
                        "description": "Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Error"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error in Requried Inputs",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Error Message",
                                            "type": "string",
                                            "example": "Unprocessable Entity"
                                        },
                                        "errors": {
                                            "description": "Errors",
                                            "type": "Object",
                                            "example": "{'email': ['الإيميل مطلوب'   ]}"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Error": {
                "title": "Error",
                "description": "Error schema",
                "properties": {
                    "title": {
                        "description": "Error Message",
                        "type": "string",
                        "example": "  error meassage "
                    },
                    "desc": {
                        "description": "Error description",
                        "type": "string",
                        "example": "Something went wrong, please try again later."
                    }
                },
                "type": "object"
            },
...
            "ForgetPasswordOtpSendRequest": {
                "title": "Forget Password Otp Send Request Schema",
                "description": "",
                "properties": {
                    "email": {
                        "title": "email",
                        "description": " email of the user ",
                        "type": "string",
                        "example": "osama.moh.almamari@gmail.com"
                    }
                },
                "type": "object"
            },
            "LoginRequest": {
                "title": "Login Request",
                "description": "LoginRequest",
                "required": [
                    "name"
                ],
                "properties": {
                    "email": {
                        "title": "email",
                        "description": "email of register account ",
                        "type": "string",
                        "example": "osama.moh.almamari@gmail.com"
                    },
                    "password": {
                        "title": "password",
                        "description": "password  ",
                        "type": "string",
                        "example": "123456"
                    }
                },
                "type": "object"
            },
            "OtpVerifyRequest": {
                "title": "OtpVerify Request",
                "description": "Otp Verify",
                "properties": {
                    "email": {
                        "title": "email",
                        "description": " email ",
                        "type": "string",
                        "example": "osama.moh.almamari@gmail.com"
                    },
                    "otp": {
                        "title": "otp",
                        "description": "otp  ",
                        "type": "string",
                        "example": "1236"
                    },
                    "password": {
                        "title": "password",
                        "description": "password  ",
                        "type": "string",
                        "example": "123456"
                    },
                    "password_confirmation": {
                        "title": "password_confirmation",
                        "description": "password_confirmation  ",
                        "type": "string",
                        "example": "123456"
                    }
                },
                "type": "object"
            },
...
            "User": {
                "title": "User Data",
                "description": "Schema for user data",
                "properties": {
                    "id": {
                        "description": "The ID of the user",
                        "type": "string",
                        "example": "01h84ym49w7byskc4wrcmnyszc"
                    },
                    "name": {
                        "description": "The name of the user",
                        "type": "string",
                        "example": "اسامة"
                    },
                    "email": {
                        "description": "The email address of the user",
                        "type": "string",
                        "example": "osama.moh.almamari@gmail.com"
                    },
                    "token": {
                        "description": "tokrn user",
                        "type": "string",
                        "example": "01h84ym49w7byskc4wrcmnyszcddddddddddddddddd"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "Authorization": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}